Refactoring, extract lighthouse outlier filter
