+++ src/modules/interface/kalman_core/outlierFilterLighthouse.h
@@ -19,24 +19,17 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
- * outlierFilter.h: Outlier rejection filter for the kalman filter
+ * Outlier rejection filter for the kalman filter
  */
 
-#ifndef __OUTLIER_FILTER_H__
-#define __OUTLIER_FILTER_H__
+#pragma once
 
 #include "stabilizer_types.h"
-
-bool outlierFilterValidateTdoaSimple(const tdoaMeasurement_t* tdoa);
-bool outlierFilterValidateTdoaSteps(const tdoaMeasurement_t* tdoa, const float error, const vector_t* jacobian, const point_t* estPos);
-bool outlierFilterValidateTdoaIntegrator(const tdoaMeasurement_t* tdoa, const float error, const uint32_t nowMs);
 
 typedef struct {
     uint32_t openingTimeMs;
     int32_t openingWindowMs;
 } OutlierFilterLhState_t;
-bool outlierFilterValidateLighthouseSweep(OutlierFilterLhState_t* this, const float distanceToBs, const float angleError, const uint32_t nowMs);
-void outlierFilterReset(OutlierFilterLhState_t* this, const uint32_t nowMs);
 
-
-#endif // __OUTLIER_FILTER_H__
+bool outlierFilterLighthouseValidateSweep(OutlierFilterLhState_t* this, const float distanceToBs, const float angleError, const uint32_t nowMs);
+void outlierFilterLighthouseReset(OutlierFilterLhState_t* this, const uint32_t nowMs);
