+++ src/modules/src/kalman_core/Kbuild
@@ -10,4 +10,5 @@ obj-y += mm_tdoa.o
 obj-y += mm_tdoa_robust.o
 obj-y += mm_tof.o
 obj-y += mm_yaw_error.o
-obj-y += outlierFilter.o
+obj-y += outlierFilterTdoa.o
+obj-y += outlierFilterLighthouse.o
